+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Control how to name log files with sequencers
 - Examples:
   - [`rotation_example`](./examples/rotation_example/main.go) - Set up a log rotator with custom naming and rotation conditions
+  - [`sequencer_example`](./examples/sequencer_example/main.go) - Set up a log rotator with a custom sequencer function
 
 **Fixed:**